Add HUD with speed display
Other changes:
- Remove redundant sprite files
- Move all spritesheets into one folder

The font is not distributed yet, I still need to figure out a few things.
Do I want to include it in the repo? If I do, do I still want to load it from a file, or give it a spot in static memory and load it from memory instead?